added Day & Night Cycle
created a sun and moon. Got them to automatically cycle throughout the day.
