--- Home Street Home/Assets/Inventory.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 35380d0d7f86c3144b03b9731b6a2e5c
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
