+++ Home Street Home/Assets/Scripts/Inventory.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 4d715cc8900b3f541a27d426adaf9bfd
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
