+++ Home Street Home/Assets/Scripts/Inventory/Inventoryui.cs
@@ -1,0 +1,15 @@
+﻿using UnityEngine;
+
+public class InventoryUI : MonoBehaviour {
+    
+
+	// Use this for initialization
+	void Start () {
+		
+	}
+	
+	// Update is called once per frame
+	void Update () {
+		
+	}
+}
