--- Home Street Home/Assets/Scripts/Inventoryui.cs
@@ -1,15 +1,0 @@
-﻿using UnityEngine;
-
-public class InventoryUI : MonoBehaviour {
-    
-
-	// Use this for initialization
-	void Start () {
-		
-	}
-	
-	// Update is called once per frame
-	void Update () {
-		
-	}
-}
