--- Home Street Home/Assets/Scripts/Inventoryui.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: c5f740d3a9ec16641988c537b2e2f49e
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
